+++ src/textual/widget.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from asyncio import Lock, wait, create_task
+from asyncio import Lock, wait, create_task, Event as AsyncEvent
 from fractions import Fraction
 from itertools import islice
 from operator import attrgetter
@@ -34,6 +34,7 @@ from . import errors, events, messages
 from ._animator import BoundAnimator, DEFAULT_EASING, Animatable, EasingFunction
 from ._arrange import DockArrangeResult, arrange
 from ._context import active_app
+from ._easing import DEFAULT_SCROLL_EASING
 from ._layout import Layout
 from ._segment_tools import align_lines
 from ._styles_cache import StylesCache
@@ -48,6 +49,7 @@ from .message import Message
 from .messages import CallbackType
 from .reactive import Reactive
 from .render import measure
+from .await_remove import AwaitRemove
 
 if TYPE_CHECKING:
     from .app import App, ComposeResult
@@ -147,6 +149,14 @@ class RenderCache(NamedTuple):
     lines: Lines
 
 
+class WidgetError(Exception):
+    """Base widget error."""
+
+
+class MountError(WidgetError):
+    """Error raised when there was a problem with the mount request."""
+
+
 @rich.repr.auto
 class Widget(DOMNode):
     """
@@ -233,6 +243,10 @@ class Widget(DOMNode):
             id=id,
             classes=self.DEFAULT_CLASSES if classes is None else classes,
         )
+
+        if self in children:
+            raise WidgetError("A widget can't be its own parent")
+
         self._add_children(*children)
 
     virtual_size = Reactive(Size(0, 0), layout=True)
@@ -374,9 +388,6 @@ class Widget(DOMNode):
         if self._scrollbar_corner is not None:
             yield self._scrollbar_corner
 
-    class MountError(Exception):
-        """Error raised when there was a problem with the mount request."""
-
     def _find_mount_point(self, spot: int | str | "Widget") -> tuple["Widget", int]:
         """Attempt to locate the point where the caller wants to mount something.
 
@@ -387,7 +398,7 @@ class Widget(DOMNode):
             tuple[Widget, int]: The parent and the location in its child list.
 
         Raises:
-            Widget.MountError: If there was an error finding where to mount a widget.
+            MountError: If there was an error finding where to mount a widget.
 
         The rules of this method are:
 
@@ -414,7 +425,7 @@ class Widget(DOMNode):
         # have a parent? There's no way we can use it as a sibling to make
         # mounting decisions if it doesn't have a parent.
         if spot.parent is None:
-            raise self.MountError(
+            raise MountError(
                 f"Unable to find relative location of {spot!r} because it has no parent"
             )
 
@@ -422,9 +433,9 @@ class Widget(DOMNode):
         # children. We should be able to go looking for the widget's
         # location amongst its parent's children.
         try:
-            return spot.parent, spot.parent.children.index(spot)
+            return cast("Widget", spot.parent), spot.parent.children.index(spot)
         except ValueError:
-            raise self.MountError(f"{spot!r} is not a child of {self!r}") from None
+            raise MountError(f"{spot!r} is not a child of {self!r}") from None
 
     def mount(
         self,
@@ -452,7 +463,7 @@ class Widget(DOMNode):
 
         # Saying you want to mount before *and* after something is an error.
         if before is not None and after is not None:
-            raise self.MountError(
+            raise MountError(
                 "Only one of `before` or `after` can be handled -- not both"
             )
 
@@ -1045,7 +1056,7 @@ class Widget(DOMNode):
         Returns:
             tuple[str, ...]: Tuple of layer names.
         """
-        for node in self.ancestors:
+        for node in self.ancestors_with_self:
             if not isinstance(node, Widget):
                 break
             if node.styles.has_rule("layers"):
@@ -1127,6 +1138,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll to a given (absolute) coordinate, optionally animating.
 
@@ -1136,6 +1148,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate to new scroll position. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the default scrolling easing function.
 
         Returns:
             bool: True if the scroll position changed, otherwise False.
@@ -1145,6 +1159,10 @@ class Widget(DOMNode):
             # TODO: configure animation speed
             if duration is None and speed is None:
                 speed = 50
+
+            if easing is None:
+                easing = DEFAULT_SCROLL_EASING
+
             if x is not None:
                 self.scroll_target_x = x
                 if x != self.scroll_x:
@@ -1153,7 +1171,7 @@ class Widget(DOMNode):
                         self.scroll_target_x,
                         speed=speed,
                         duration=duration,
-                        easing="out_cubic",
+                        easing=easing,
                     )
                     scrolled_x = True
             if y is not None:
@@ -1164,7 +1182,7 @@ class Widget(DOMNode):
                         self.scroll_target_y,
                         speed=speed,
                         duration=duration,
-                        easing="out_cubic",
+                        easing=easing,
                     )
                     scrolled_y = True
 
@@ -1188,6 +1206,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll relative to current position.
 
@@ -1197,6 +1216,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate to new scroll position. Defaults to False.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if the scroll position changed, otherwise False.
@@ -1207,6 +1228,7 @@ class Widget(DOMNode):
             animate=animate,
             speed=speed,
             duration=duration,
+            easing=easing,
         )
 
     def scroll_home(
@@ -1215,6 +1237,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll to home position.
 
@@ -1222,13 +1245,17 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
         """
         if speed is None and duration is None:
             duration = 1.0
-        return self.scroll_to(0, 0, animate=animate, speed=speed, duration=duration)
+        return self.scroll_to(
+            0, 0, animate=animate, speed=speed, duration=duration, easing=easing
+        )
 
     def scroll_end(
         self,
@@ -1236,6 +1263,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll to the end of the container.
 
@@ -1243,6 +1271,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
@@ -1251,7 +1281,12 @@ class Widget(DOMNode):
         if speed is None and duration is None:
             duration = 1.0
         return self.scroll_to(
-            0, self.max_scroll_y, animate=animate, speed=speed, duration=duration
+            0,
+            self.max_scroll_y,
+            animate=animate,
+            speed=speed,
+            duration=duration,
+            easing=easing,
         )
 
     def scroll_left(
@@ -1260,6 +1295,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one cell left.
 
@@ -1267,13 +1303,19 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
 
         """
         return self.scroll_to(
-            x=self.scroll_target_x - 1, animate=animate, speed=speed, duration=duration
+            x=self.scroll_target_x - 1,
+            animate=animate,
+            speed=speed,
+            duration=duration,
+            easing=easing,
         )
 
     def scroll_right(
@@ -1282,6 +1324,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll on cell right.
 
@@ -1289,13 +1332,19 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
 
         """
         return self.scroll_to(
-            x=self.scroll_target_x + 1, animate=animate, speed=speed, duration=duration
+            x=self.scroll_target_x + 1,
+            animate=animate,
+            speed=speed,
+            duration=duration,
+            easing=easing,
         )
 
     def scroll_down(
@@ -1304,6 +1353,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one line down.
 
@@ -1311,13 +1361,19 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
 
         """
         return self.scroll_to(
-            y=self.scroll_target_y + 1, animate=animate, speed=speed, duration=duration
+            y=self.scroll_target_y + 1,
+            animate=animate,
+            speed=speed,
+            duration=duration,
+            easing=easing,
         )
 
     def scroll_up(
@@ -1326,6 +1382,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one line up.
 
@@ -1333,13 +1390,19 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
 
         """
         return self.scroll_to(
-            y=self.scroll_target_y - 1, animate=animate, speed=speed, duration=duration
+            y=self.scroll_target_y - 1,
+            animate=animate,
+            speed=speed,
+            duration=duration,
+            easing=easing,
         )
 
     def scroll_page_up(
@@ -1348,6 +1411,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one page up.
 
@@ -1355,6 +1419,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
@@ -1365,6 +1431,7 @@ class Widget(DOMNode):
             animate=animate,
             speed=speed,
             duration=duration,
+            easing=easing,
         )
 
     def scroll_page_down(
@@ -1373,6 +1440,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one page down.
 
@@ -1380,6 +1448,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
@@ -1390,6 +1460,7 @@ class Widget(DOMNode):
             animate=animate,
             speed=speed,
             duration=duration,
+            easing=easing,
         )
 
     def scroll_page_left(
@@ -1398,6 +1469,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one page left.
 
@@ -1405,6 +1477,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
@@ -1417,6 +1491,7 @@ class Widget(DOMNode):
             animate=animate,
             speed=speed,
             duration=duration,
+            easing=easing,
         )
 
     def scroll_page_right(
@@ -1425,6 +1500,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
     ) -> bool:
         """Scroll one page right.
 
@@ -1432,6 +1508,8 @@ class Widget(DOMNode):
             animate (bool, optional): Animate scroll. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
 
         Returns:
             bool: True if any scrolling was done.
@@ -1444,6 +1522,7 @@ class Widget(DOMNode):
             animate=animate,
             speed=speed,
             duration=duration,
+            easing=easing,
         )
 
     def scroll_to_widget(
@@ -1453,6 +1532,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
         top: bool = False,
     ) -> bool:
         """Scroll scrolling to bring a widget in to view.
@@ -1462,6 +1542,9 @@ class Widget(DOMNode):
             animate (bool, optional): True to animate, or False to jump. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
+            top (bool, optional): Scroll widget to top of container. Defaults to False.
 
         Returns:
             bool: True if any scrolling has occurred in any descendant, otherwise False.
@@ -1480,6 +1563,7 @@ class Widget(DOMNode):
                 speed=speed,
                 duration=duration,
                 top=top,
+                easing=easing,
             )
             if scroll_offset:
                 scrolled = True
@@ -1506,6 +1590,7 @@ class Widget(DOMNode):
         animate: bool = True,
         speed: float | None = None,
         duration: float | None = None,
+        easing: EasingFunction | str | None = None,
         top: bool = False,
     ) -> Offset:
         """Scrolls a given region in to view, if required.
@@ -1519,6 +1604,8 @@ class Widget(DOMNode):
             animate (bool, optional): True to animate, or False to jump. Defaults to True.
             speed (float | None, optional): Speed of scroll if animate is True. Or None to use duration.
             duration (float | None, optional): Duration of animation, if animate is True and speed is None.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
             top (bool, optional): Scroll region to top of container. Defaults to False.
 
         Returns:
@@ -1546,6 +1633,7 @@ class Widget(DOMNode):
                 animate=animate if (abs(delta_y) > 1 or delta_x) else False,
                 speed=speed,
                 duration=duration,
+                easing=easing,
             )
         return delta
 
@@ -1556,6 +1644,7 @@ class Widget(DOMNode):
         speed: float | None = None,
         duration: float | None = None,
         top: bool = False,
+        easing: EasingFunction | str | None = None,
     ) -> None:
         """Scroll the container to make this widget visible.
 
@@ -1564,16 +1653,19 @@ class Widget(DOMNode):
             speed (float | None, optional): _description_. Defaults to None.
             duration (float | None, optional): _description_. Defaults to None.
             top (bool, optional): Scroll to top of container. Defaults to False.
+            easing (EasingFunction | str | None, optional): An easing method for the scrolling animation. Defaults to "None",
+                which will result in Textual choosing the configured default scrolling easing function.
         """
         parent = self.parent
         if isinstance(parent, Widget):
-            self.call_later(
+            self.call_after_refresh(
                 parent.scroll_to_widget,
                 self,
                 animate=animate,
                 speed=speed,
                 duration=duration,
                 top=top,
+                easing=easing,
             )
 
     def __init_subclass__(
@@ -1899,9 +1991,21 @@ class Widget(DOMNode):
 
         self.check_idle()
 
-    def remove(self) -> None:
-        """Remove the Widget from the DOM (effectively deleting it)"""
-        self.app.post_message_no_wait(events.Remove(self, widget=self))
+    def remove(self) -> AwaitRemove:
+        """Remove the Widget from the DOM (effectively deleting it)
+
+        Returns:
+            AwaitRemove: An awaitable object that waits for the widget to be removed.
+        """
+        prune_finished_event = AsyncEvent()
+        self.app.post_message_no_wait(
+            events.Prune(
+                self,
+                widgets=self.app._detach_from_dom([self]),
+                finished_flag=prune_finished_event,
+            )
+        )
+        return AwaitRemove(prune_finished_event)
 
     def render(self) -> RenderableType:
         """Get renderable for widget.
@@ -2050,14 +2154,14 @@ class Widget(DOMNode):
         self.mouse_over = True
 
     def _on_focus(self, event: events.Focus) -> None:
-        for node in self.ancestors:
+        for node in self.ancestors_with_self:
             if node._has_focus_within:
                 self.app.update_styles(node)
         self.has_focus = True
         self.refresh()
 
     def _on_blur(self, event: events.Blur) -> None:
-        if any(node._has_focus_within for node in self.ancestors):
+        if any(node._has_focus_within for node in self.ancestors_with_self):
             self.app.update_styles(self)
         self.has_focus = False
         self.refresh()