+++ src/textual/widgets/_data_table.py
@@ -56,7 +56,7 @@ class Column:
 
     @property
     def render_width(self) -> int:
-        """Width in cells, required to render a column."""
+        """int: Width in cells, required to render a column."""
         # +2 is to account for space padding either side of the cell
         if self.auto_width:
             return self.content_width + 2
@@ -88,22 +88,38 @@ class Coord(NamedTuple):
     column: int
 
     def left(self) -> Coord:
-        """Get coordinate to the left."""
+        """Get coordinate to the left.
+
+        Returns:
+            Coord: The coordinate.
+        """
         row, column = self
         return Coord(row, column - 1)
 
     def right(self) -> Coord:
-        """Get coordinate to the right."""
+        """Get coordinate to the right.
+
+        Returns:
+            Coord: The coordinate.
+        """
         row, column = self
         return Coord(row, column + 1)
 
     def up(self) -> Coord:
-        """Get coordinate above."""
+        """Get coordinate above.
+
+        Returns:
+            Coord: The coordinate.
+        """
         row, column = self
         return Coord(row - 1, column)
 
     def down(self) -> Coord:
-        """Get coordinate below."""
+        """Get coordinate below.
+
+        Returns:
+            Coord: The coordinate.
+        """
         row, column = self
         return Coord(row + 1, column)
 
@@ -275,6 +291,7 @@ class DataTable(ScrollView, Generic[CellType], can_focus=True):
                 content_width = measure(self.app.console, renderable, 1)
                 column.content_width = max(column.content_width, content_width)
 
+        self._clear_caches()
         total_width = sum(column.render_width for column in self.columns)
         header_height = self.header_height if self.show_header else 0
         self.virtual_size = Size(
@@ -294,6 +311,21 @@ class DataTable(ScrollView, Generic[CellType], can_focus=True):
             y += self.header_height
         cell_region = Region(x, y, width, height)
         return cell_region
+
+    def clear(self) -> None:
+        """Clear the table.
+
+        Args:
+            columns (bool, optional): Also clear the columns. Defaults to False.
+        """
+        self.row_count = 0
+        self._clear_caches()
+        self._y_offsets.clear()
+        self.data.clear()
+        self.rows.clear()
+        self._line_no = 0
+        self._require_update_dimensions = True
+        self.refresh()
 
     def add_columns(self, *labels: TextType) -> None:
         """Add a number of columns.
@@ -369,6 +401,7 @@ class DataTable(ScrollView, Generic[CellType], can_focus=True):
             new_rows = self._new_rows.copy()
             self._new_rows.clear()
             self._update_dimensions(new_rows)
+            self.refresh()
 
     def refresh_cell(self, row_index: int, column_index: int) -> None:
         """Refresh a cell.